requirements.txt
mysql-connector-python
python-dotenv
pyyaml
pytz
#tabulate
babel
openpyxl